requirements.txt
pandas==1.0.3
transformers==2.10.0
pydot==1.4.1
google-cloud-storage==1.28.1
gcsfs==0.6.2
